refactor: store.js to ts - add package @types/redux-logger - change store.js to store.ts - type RootState
