+++ package.json
@@ -10,6 +10,7 @@
     "@types/node": "^18.11.18",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
+    "@types/redux-logger": "^3.0.9",
     "firebase": "9.6.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
