+++ src/store/cart/cart.selector.ts
@@ -1,7 +1,8 @@
+import { RootState } from "./../store";
 import { CartState } from "./cart.reducer";
 import { createSelector } from "reselect";
 
-const selectCartReducer = (state): CartState => state.cart;
+const selectCartReducer = (state: RootState): CartState => state.cart;
 
 export const selectCartItems = createSelector(
   [selectCartReducer],
