+++ src/store/categories/category.selector.ts
@@ -1,8 +1,10 @@
+import { RootState } from "./../store";
 import { CategoryTable } from "./category.types";
 import { createSelector } from "reselect";
 import { CategoriesState } from "./category.reducer";
 
-const selectCategoryReducer = (state): CategoriesState => state.categories;
+const selectCategoryReducer = (state: RootState): CategoriesState =>
+  state.categories;
 
 export const selectCategories = createSelector(
   [selectCategoryReducer],
