+++ src/store/user/user.selector.ts
@@ -1,7 +1,8 @@
+import { RootState } from "./../store";
 import { UserState } from "./user.reducer";
 import { createSelector } from "reselect";
 
-export const selectUserReducer = (state): UserState => state.user;
+export const selectUserReducer = (state: RootState): UserState => state.user;
 
 export const selectCurrentUser = createSelector(
   selectUserReducer,
